MakeLazyComponent now returns a component with the spinner props
Switch AdvancedDyanmicTextureGeneralProperties to use MakeLazyComponent as part of the implementation
